pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "epilepsure"
version = "2.0.0"
dependencies = [
    "openenv-core",
    "numpy",
    "opencv-python",
    "fastapi",
    "uvicorn",
    "pydantic",
    "openai",
    "Pillow",
]
